Updated Search and Filter Functionality
